server/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
python-socketio==5.8.0
python-engineio==4.5.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
